fix: continue macOS port - Windows API stubs and type fixes
- Add extensive Windows API compatibility layer in Kernel.h
- Add missing types: COLORREF, LPCTSTR, BITMAP, BITMAPINFOHEADER, etc.
- Fix CString operator== for const char* comparisons
- Fix VoronoiGenerator.h LOG macros (/##/ → empty definitions)
- Fix CCAtomTable.cpp pointer/int type mismatches
- Fix CCPrimitive.cpp function pointer to void* cast
- Remove Windows GDI files from build (DIB, CGBitmap, CGFont, GDI, Regions, CGResourceFile)
- Fix CDataFile.cpp, CIDTable.cpp, CStackBase.cpp, CString.cpp
- Fix Path.cpp, Math.cpp, Utilities.cpp Windows-specific sections
- Add DirectXUtilCompat.h clipboard function guards
- Progress: compilation errors reduced to linker errors

+++ Alchemy/Kernel/CIDTable.cpp
@@ -168,7 +168,7 @@ ALERROR CIDTable::LoadHandler (CUnarchiver *pUnarchiver)
 			throw CException(ERR_FAIL);
 #endif
 
-		CDictionary::SetEntry(i, iKey, (int)pValue);
+		CDictionary::SetEntry(i, iKey, (intptr_t)pValue);
 		}
 
 	return NOERROR;
@@ -273,7 +273,7 @@ ALERROR CIDTable::ReplaceEntry (int iKey, CObject *pValue, bool bAdd, CObject **
 
 	//	Let the dictionary code do the actual adding
 
-	if (error = CDictionary::ReplaceEntry(iKey, (int)pValue, bAdd, &bAdded, &iOldValue))
+	if (error = CDictionary::ReplaceEntry(iKey, (intptr_t)pValue, bAdd, &bAdded, &iOldValue))
 		return error;
 
 	//	If we added a new object, then there is no old value
@@ -379,5 +379,5 @@ void CIDTable::SetValue (int iEntry, CObject *pValue, CObject **retpOldValue)
 	if (retpOldValue)
 		*retpOldValue = (CObject *)iValue;
 
-	CDictionary::SetEntry(iEntry, iKey, (int)pValue);
+	CDictionary::SetEntry(iEntry, iKey, (intptr_t)pValue);
 	}
