+++ Mammoth/TSE/CNavigationPath.cpp
@@ -341,7 +341,7 @@ CString CNavigationPath::DebugDescribe (CSpaceObject *pObj, CNavigationPath *pNa
 		}
 	catch (...)
 		{
-		return strPatternSubst(CONSTLIT("%x [invalid]"), (DWORD)pNavPath);
+		return strPatternSubst(CONSTLIT("%x [invalid]"), (DWORD)(uintptr_t)pNavPath);
 		}
 
 	//	Get the objects
